Changed StackArray to use floats instead of longs to store its elements. And changed StackVM to use floats instead of longs when interacting with the stack. This is probably a very very bad idea in the long run, but I want to see how far I get with this approach.

+++ c++/UziFirmware/StackArray.cpp
@@ -1,18 +1,18 @@
 #include "StackArray.h"
 
 
-void StackArray::push(long element) {
+void StackArray::push(float element) {
 	_elements[_pointer] = element;
 	_pointer++;
 }
 
-long StackArray::pop(void) {
+float StackArray::pop(void) {
 	_pointer--;
-	long result = _elements[_pointer];	
+	float result = _elements[_pointer];	
 	return result;
 }
 
-long StackArray::top(void) {
+float StackArray::top(void) {
 	return _elements[_pointer - 1];
 }
