+++ c++/UziFirmware/StackVM.cpp
@@ -36,8 +36,8 @@ void StackVM::executeBytecode(unsigned char bytecode) {
 	unsigned char key = bytecode & 0xF0;
 	unsigned char value = bytecode & 0x0F;
 	
-	long a;
-	long b;
+	float a;
+	float b;
 
 	switch (key) {
 		case 0x00:// pushLit
@@ -91,8 +91,8 @@ void StackVM::executeBytecode(unsigned char bytecode) {
 
 void StackVM::executePrimitive(unsigned char primitiveIndex) {
 
-	long pop1;
-	long pop2;
+	float pop1;
+	float pop2;
 
 	switch(primitiveIndex) {
 		case 0x00:// getValue
@@ -102,7 +102,7 @@ void StackVM::executePrimitive(unsigned char primitiveIndex) {
 		case 0x01:// setValue
 			pop1 = _stack->pop();
 			pop2 = _stack->pop(); 
-			_pe->setValue((unsigned int)pop2, (unsigned short)pop1);
+			_pe->setValue((unsigned int)pop2, pop1);
 			break;
 		case 0x02:// getMode
 			pop1 = _stack->pop();
